requirements.txt
Django==4.1.5
Pillow==9.3.0
pip==22.3.1
wheel==0.38.4
django-axes==5.40.1
setuptools==65.6.3
asgiref==3.6.0
sqlparse==0.4.3
tablib==3.3.0
django-crispy-forms==1.12.0
tzdata==2022.7
phonenumbers==8.13.3
django-password-validators==1.7.0
django-sslserver==0.22
django-phonenumber-field==7.0.2
django-import-export==3.0.2